+++ CHANGELOG.md
@@ -7,6 +7,20 @@ If it changes significantly or very much, the first digit will move.
 
 ***
 
+## [Beta 1.2](https://gitlab.com/Avespoir_Project/Avespoir/-/tree/Beta-1.2) - 2020-05-19
+
+### Added
+
+* One-time Authentication
+
+* Moderator Commands
+  * DBLogChannelAdd
+  * DBLogChannelChange
+
+### Removed
+
+* Botowner Command: DBChannelAdd
+
 ## [Beta 1.1.1](https://gitlab.com/Avespoir_Project/Avespoir/-/tree/Beta-1.1.1) - 2020-03-26
 
 ### Fixed